+++ src/JavaScriptActionBuilder/PackageManager/PackageJsonSchema.ts
